requirements.txt
ansicon==1.89.0
blessed==1.20.0
certifi==2024.8.30
cloudinary==1.41.0
colorama==0.4.6
editor==1.6.6
filetype==1.2.0
fuzzywuzzy==0.18.0
ImageHash==4.3.1
iniconfig==2.0.0
inquirer==3.4.0
jinxed==1.3.0
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pynput==1.7.7
pytest==8.3.3
python-dotenv==1.0.1
PyWavelets==1.7.0
RapidFuzz==3.10.1
readchar==4.2.1
runs==1.2.2
scipy==1.14.1
setuptools==75.5.0
six==1.16.0
urllib3==2.2.3
wcwidth==0.2.13
xmod==1.8.1
